Add ability to clone an existing db

+++ machines/server.ts
@@ -2,7 +2,6 @@ import * as Y from "yjs"
 
 const isServerSync = true
 export function listen({ doc, serverConfig }) {
-  console.log(`ydoc listen`)
   const requests = doc.getArray(`requests`)
   requests.observe(async (event) => {
     const inserted = event.changes.delta.find(
@@ -14,25 +13,38 @@ export function listen({ doc, serverConfig }) {
     const itemArray = retain?.retain || 0
     const state = event.target.get(itemArray)
     if (state.done !== true) {
-      const mutatorFunc = await serverConfig.mutators[state.mutator]({
-        state,
-        doc,
-      })
-      doc.transact(() => {
-        state.response = mutatorFunc()
+      if (
+        serverConfig.mutators.hasOwnProperty(state.mutator) &&
+        serverConfig.mutators[state.mutator] instanceof Function
+      ) {
+        const mutatorFunc = await serverConfig.mutators[state.mutator]({
+          state,
+          doc,
+        })
+        doc.transact(() => {
+          state.response = mutatorFunc()
 
-        if (typeof state.response?.error !== `undefined`) {
+          if (typeof state.response?.error !== `undefined`) {
+            state.error = true
+          } else {
+            state.error = false
+          }
+
+          state.value = `responded`
+          state.done = true
+          state.serverResponded = new Date().toJSON()
+          requests.delete(itemArray, 1)
+          requests.insert(itemArray, [state])
+        })
+      } else {
+        console.log(`bad request`)
+        doc.transact(() => {
           state.error = true
-        } else {
-          state.error = false
-        }
-
-        state.value = `responded`
-        state.done = true
-        state.serverResponded = new Date().toJSON()
-        requests.delete(itemArray, 1)
-        requests.insert(itemArray, [state])
-      })
+          state.response = {
+            error: `A mutator by that name does not exist`,
+          }
+        })
+      }
     }
   })
 }
